+++ sclite/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-__version__ = '0.7.0a0'
+__version__ = '0.8.0a0'
 
 from .bundles import (
     REVIEW_BUNDLE_MANIFEST_FILE,